+++ example/lib/main.dart
@@ -13,22 +13,15 @@ class MyApp extends AppMVC {
 
   @override
   AppState createState() => AppState(
-        title: 'Flutter Demo',
         theme: ThemeData(
           // This is the theme of your application.
-          //
-          // Try running your application with "flutter run". You'll see the
-          // application has a blue toolbar. Then, without quitting the app, try
-          // changing the primarySwatch below to Colors.green and then invoke
-          // "hot reload" (press "r" in the console where you ran "flutter run",
-          // or simply save your changes to "hot reload" in a Flutter IDE).
-          // Notice that the counter didn't reset back to zero; the application
-          // is not restarted.
           primarySwatch: Colors.blue,
         ),
+        // This is the app's navigation
         routerDelegate: AppRouterDelegate(routes: {
           '/': (_) => MyHomePage(),
         }),
+        // This is the app's localization
         supportedLocales: I10n.supportedLocales,
         localizationsDelegates: [
           I10nDelegate(),
@@ -36,6 +29,7 @@ class MyApp extends AppMVC {
           GlobalCupertinoLocalizations.delegate,
           GlobalMaterialLocalizations.delegate,
         ],
+        // These are flags used during development
 //    debugPaintSizeEnabled: true,
         debugShowCheckedModeBanner: false,
       );